backend/app/tenancy/tests/test_isolation.py
import pytest
from django.urls import reverse
from rest_framework import status
from rest_framework.test import APIClient

from app.accounts.models import User
from app.tenancy.models import Tenant
from app.connections.models import Connection # A sample tenant-scoped resource

# Mark all tests in this module as Django DB tests
pytestmark = pytest.mark.django_db

@pytest.fixture
def tenants_and_users():
    """
    Fixture to create two separate tenants and a user for each.
    """
    # Tenant A setup
    tenant_a = Tenant.objects.create(name="Tenant A")
    user_a = User.objects.create_user(username="user_a", password="password123")
    user_a.tenants.add(tenant_a) # Assuming a ManyToManyField on User or through Membership

    # Tenant B setup
    tenant_b = Tenant.objects.create(name="Tenant B")
    user_b = User.objects.create_user(username="user_b", password="password123")
    user_b.tenants.add(tenant_b)

    return tenant_a, user_a, tenant_b, user_b

def test_user_cannot_access_other_tenant_resources(tenants_and_users):
    """
    This is a critical test to ensure tenant isolation.
    It verifies that a user logged into Tenant A cannot access resources
    that belong to Tenant B.
    """
    tenant_a, user_a, tenant_b, user_b = tenants_and_users

    # Create a resource that belongs to Tenant B
    connection_b = Connection.objects.create(
        tenant=tenant_b,
        name="Tenant B's Database",
        driver=Connection.Driver.POSTGRES,
        created_by=user_b
    )

    # --- Test Logic ---
    # 1. Authenticate as User A
    client = APIClient()
    client.login(username="user_a", password="password123")

    # 2. Set the tenant context for User A's requests
    # In a real app, this would be handled by middleware, perhaps via a header.
    # For testing, we can simulate it.
    client.defaults['HTTP_X_TENANT_ID'] = tenant_a.id

    # 3. Try to access the resource from Tenant B
    # This assumes a DRF detail view for the Connection model exists.
    # url = reverse('connection-detail', kwargs={'pk': connection_b.pk})
    url = f"/api/v1/connections/{connection_b.pk}/" # Placeholder URL

    # 4. Assert that the request is denied
    # The expected outcome is a 404 Not Found, as if the resource
    # doesn't exist from User A's perspective.
    # response = client.get(url)
    # assert response.status_code == status.HTTP_404_NOT_FOUND

    # --- Placeholder Assertion ---
    # Since the API doesn't exist yet, we'll just assert True.
    # This structure should be filled in once the API endpoints are built.
    assert True, "This test should be implemented with a real API call."

    # Also test the reverse: listing resources for Tenant A should not show Tenant B's items
    # list_url = reverse('connection-list')
    # response = client.get(list_url)
    # assert response.status_code == status.HTTP_200_OK
    # assert connection_b.name not in str(response.content)
    assert True, "A list-view test should also be implemented."
